Updated to version 1.3
Modified design to work on mobile screen sizes.
Added log service for unexpected errors.
Made final adjusments for production.

+++ src/components/Label/index.jsx
@@ -24,12 +24,17 @@ function Label({ label, onDelete, onUpdate, hovering }) {
 
   async function handleDelete(e) {
     e.stopPropagation();
-    setDeleting(true);
+    const confirmation = window.confirm(
+      "Are you sure? \nDeleting this label will remove it from all corresponding tasks."
+    );
 
-    try {
-      await onDelete(label._id).unwrap();
-    } catch (ex) {
-      setDeleting(false);
+    if (confirmation) {
+      setDeleting(true);
+      try {
+        await onDelete(label._id).unwrap();
+      } catch (ex) {
+        setDeleting(false);
+      }
     }
   }
